Share Blackwell safe_dot helper
