+++ fla/ops/utils/op.py
@@ -11,7 +11,7 @@ import triton
 import triton.language as tl
 import triton.language.extra.libdevice as tldevice
 
-from fla.utils import IS_GATHER_SUPPORTED
+from fla.utils import IS_GATHER_SUPPORTED, IS_NVIDIA_BLACKWELL
 
 if os.environ.get('FLA_USE_FAST_OPS', '0') == '1':
     @triton.jit
@@ -35,6 +35,33 @@ else:
     def log2(x): return tl.log2(x.to(tl.float32))
     @triton.jit
     def tanh(x): return tldevice.tanh(x.to(tl.float32))
+
+
+if IS_NVIDIA_BLACKWELL:
+    """
+    Compute tl.dot with SM100 workaround.
+
+    On SM100 (Blackwell) GPUs, wraps the result in inline assembly to prevent
+    the TritonGPUHoistTMEMAlloc pass from incorrectly fusing add and dot operations.
+    See: https://github.com/fla-org/flash-linear-attention/issues/638
+
+    TODO: Remove this workaround once the Triton compiler bug is fixed.
+    Track upstream issue at: https://github.com/triton-lang/triton/issues/8695
+    """
+    @triton.jit
+    def safe_dot(a, b):
+        return tl.inline_asm_elementwise(
+            asm="mov.f32 $0, $1;",
+            constraints="=r,r",
+            args=[tl.dot(a, b)],
+            dtype=tl.float32,
+            is_pure=True,
+            pack=1,
+        )
+else:
+    @triton.jit
+    def safe_dot(a, b):
+        return tl.dot(a, b)
 
 
 if not IS_GATHER_SUPPORTED:
